+++ packages/core/src/console/kernel.ts
@@ -84,17 +84,6 @@ export class ConsoleKernel implements ConsoleKernelContract {
   }
 
   /**
-   * Returns the list of command class names to exclude when loading all
-   * commands into the console application. This is useful for base
-   * commands that provide shared functionality to other commands.
-   *
-   * @returns {Array}
-   */
-  excludedCommands (): string[] {
-    return []
-  }
-
-  /**
    * Register the console commands to the application.
    */
   async commands (): Promise<void> {
@@ -118,9 +107,6 @@ export class ConsoleKernel implements ConsoleKernelContract {
       .map((commandFile: string) => {
         return this.resolve(commandFile) as any // TODO fix this typing issue in @supercharge/collections
       })
-      .filter((command: Command) => {
-        return command instanceof Command && this.isNotExcluded(command)
-      })
       .forEach((command: Command) => {
         this.craft().add(command)
       })
@@ -137,32 +123,6 @@ export class ConsoleKernel implements ConsoleKernelContract {
     const CommandCtor = esmRequire(commandFile)
 
     return new CommandCtor()
-  }
-
-  /**
-   * Determine whether the given `command` should not be excluded
-   * from loading it to the list of available commands.
-   *
-   * @param {Command} command
-   *
-   * @returns {Boolean}
-   */
-  isNotExcluded (command: Command): boolean {
-    return !this.shouldExclude(command)
-  }
-
-  /**
-   * Determine whether the given `command` should be excluded
-   * from loading it to the list of available commands.
-   *
-   * @param {Command} command
-   *
-   * @returns {Boolean}
-   */
-  shouldExclude (command: Command): boolean {
-    return this.excludedCommands().includes(
-      command.constructor.name
-    )
   }
 
   /**