+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## [2.0.0-alpha.8](https://github.com/supercharge/framework/compare/v2.0.0-alpha.6...v2.0.0-alpha.8) - 2021-12-xx
+## [2.0.0-alpha.9](https://github.com/supercharge/framework/compare/v2.0.0-alpha.8...v2.0.0-alpha.9) - 2022-xx-xx
+
+### Updated
+- `@supercharge/container `
+  - refine container contracts: retrieve the resolved type when using a class in `make`
+- `@supercharge/http`
+  - refine input bag contracts: detect whether the `get` method retrieves a default value
+
+
+## [2.0.0-alpha.8](https://github.com/supercharge/framework/compare/v2.0.0-alpha.7...v2.0.0-alpha.8) - 2022-01-15
 
 ### Added
 - `@supercharge/contracts`